requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
langchain-text-splitters>=0.1.0
langchain-huggingface>=0.0.3
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
PyMuPDF>=1.23.0
python-docx>=0.8.11
pypdf2>=3.0.0
openpyxl>=3.1.0
uuid
